services/event-fusion-engine/requirements.txt
aiohttp==3.9.5
orjson==3.10.7
pydantic==2.8.2
redis==5.0.7
google-cloud-pubsub==2.23.1
